+++ engine/src/main/java/net/jqwik/engine/DefaultJqwikConfiguration.java
@@ -17,7 +17,7 @@ public class DefaultJqwikConfiguration implements JqwikConfiguration {
 	private TestEngineConfiguration testEngineConfiguration = null;
 
 	DefaultJqwikConfiguration(ConfigurationParameters configurationParameters) {
-		this(JqwikProperties.loadWithBackwardsCompatibility(configurationParameters));
+		this(JqwikProperties.load(configurationParameters));
 	}
 
 	private DefaultJqwikConfiguration(JqwikProperties properties) {